data_monitor_src: move from gui/ to model/
