+++ gui/src/data_monitor.cpp
@@ -3,9 +3,6 @@
 #include <fstream>
 #include <sstream>
 
-#include "data_monitor.h"
-#include "data_monitor_src.h"
-
 #include <wx/app.h>
 #include <wx/button.h>
 #include <wx/choice.h>
@@ -28,12 +25,13 @@
 #include "androidUTIL.h"
 #endif
 
+#include "model/data_monitor_src.h"
 #include "model/filters_on_disk.h"
 #include "model/navmsg_filter.h"
 #include "model/nmea_log.h"
 #include "model/gui.h"
 
-#include "data_monitor_src.h"
+#include "data_monitor.h"
 #include "svg_icons.h"
 #include "tty_scroll.h"
 #include "filter_dlg.h"
